Prestige Elm Park Channasandra Bangalore
Prestige Elm Park is a recent residential project by Prestige Group. The project is known for its contemporary design and careful planning. This ultra-modern residential project is coming up at a great location in Bengaluru.

The houses are designed with great architectural skill and built to look great. It has lush green surroundings and is done with chic fittings to give a smart look to the homes. The balconies have been made spacious for the people to enjoy the surroundings of these houses. Prestige Elm Park Bangalore provides an exclusive floor plan for 1 BHK, 2 BHK, and 3 BHK luxury residences. 

Community amenities of this place include a clubhouse, a meditation center, sports facility, a children's play area, a swimming pool, a gymnasium, landscaped garden, open space, intercom facility, power back up, 24-hour water supply, car parking, Multipurpose hall, video security and an elevator. The community facilities have been designed in such a way as to provide a great deal of convenience and comfort to the residents of this project. Greater emphasis has been placed on the security of this complex with security cameras, alarm provision, and an address system.

Location This project is located in Channasandra Whitefield Bangalore. Prestige Elm Park Whitefield is a great location that is very well connected to the rest of Bangalore. The best restaurants, hotels, shopping malls, schools, institutes, hospitals, parks, etc are close to this place. This place is also very well connected through public transport to various business places in Bengaluru. The project is well-funded by financial institutions and banks to provide you with smooth payment plans.

For More Details
Visit us - https://bit.ly/3fZP0DP